requirements.txt
Amino.py==1.2.8